pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imagic"
version = "0.1.0"
description = "Professional-grade photo editing orchestrator with AI-powered culling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.5.0",
    "SQLAlchemy>=2.0",
    "PyYAML>=6.0",
    "Pillow>=10.0",
    "numpy>=1.24",
    "rawpy>=0.19",
    "scipy>=1.11",
    "imagehash>=4.3",
]

[project.optional-dependencies]
ai = [
    "torch>=2.0",
    "transformers>=4.30",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mypy>=1.5",
    "ruff>=0.1",
]

[project.scripts]
imagic = "imagic.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
